Update vendored mujs -> 1.3.9
Part of updating mupdf and its dependencies to 1.28.0-rc2 (mujs is a mupdf
thirdparty submodule, commit e892c9f). Source-only overwrite of the .c/.h
files (+ Makefile). The new upstream UnicodeData.txt / SpecialCasing.txt are
genucd.py inputs, not compiled source (utfdata.h is unchanged), so they are
not vendored. No /* SumatraPDF local edits in this dependency.

+++ ext/mujs/jsvalue.c
@@ -170,7 +170,7 @@ const char *js_itoa(char *out, int v)
 	unsigned int a;
 	int i = 0;
 	if (v < 0) {
-		a = -v;
+		a = -(unsigned)v; /* cast to avoid -INT_MIN signed overflow UB */
 		*s++ = '-';
 	} else {
 		a = v;
